Make ETS with lookup the de facto ETS implementation
