+++ bench/list_handlers_for_event.ex
@@ -40,9 +40,7 @@ defmodule Mix.Tasks.Bench.ListHandlersForEvent do
 
   @impls %{
     "Agent" => Events.Impl.Agent,
-    "ETS" => Events.Impl.Ets,
-    "ETS with cache" => Events.Impl.EtsCached,
-    "ETS with lookup" => Events.Impl.EtsLookup
+    "ETS" => Events.Impl.Ets
   }
 
   def run(argv) do
