+++ config/config.exs
@@ -1,4 +1,1 @@
 use Mix.Config
-
-config :events,
-  impl: Events.Impl.EtsLookup
